Init Dwarf

+++ code/datums/map_config.dm
@@ -5,7 +5,7 @@
 
 /datum/map_config
 	// Metadata
-	var/config_filename = "_maps/metastation.json"
+	var/config_filename = "_maps/constructionstation.json"
 	var/defaulted = TRUE  // set to FALSE by LoadConfig() succeeding
 	// Config from maps.txt
 	var/config_max_users = 0
@@ -14,68 +14,27 @@
 	var/votable = FALSE
 
 	// Config actually from the JSON - should default to Meta
-	var/map_name = "Meta Station"
-	var/map_path = "map_files/MetaStation"
-	var/map_file = "MetaStation.dmm"
+	var/map_name = "Fortress"
+	var/map_path = "map_files/Fortress"
+	var/map_file = "Fortress.dmm"
 
 	var/traits = null
-	var/space_ruin_levels = 7
-	var/space_empty_levels = 1
-
-	var/minetype = "lavaland"
-
-	var/allow_custom_shuttles = TRUE
-	var/shuttles = list(
-		"cargo" = "cargo_box",
-		"ferry" = "ferry_fancy",
-		"whiteship" = "whiteship_box",
-		"emergency" = "emergency_box")
 
 	/// Dictionary of job sub-typepath to template changes dictionary
 	var/job_changes = list()
 
-/**
- * Proc that simply loads the default map config, which should always be functional.
- */
-/proc/load_default_map_config()
-	return new /datum/map_config
-
-
-/**
- * Proc handling the loading of map configs. Will return the default map config using [/proc/load_default_map_config] if the loading of said file fails for any reason whatsoever, so we always have a working map for the server to run.
- * Arguments:
- * * filename - Name of the config file for the map we want to load. The .json file extension is added during the proc, so do not specify filenames with the extension.
- * * directory - Name of the directory containing our .json - Must be in MAP_DIRECTORY_WHITELIST. We default this to MAP_DIRECTORY_MAPS as it will likely be the most common usecase. If no filename is set, we ignore this.
- * * error_if_missing - Bool that says whether failing to load the config for the map will be logged in log_world or not as it's passed to LoadConfig().
- *
- * Returns the config for the map to load.
- */
-/proc/load_map_config(filename = null, directory = null, error_if_missing = TRUE)
-	var/datum/map_config/config = load_default_map_config()
-
-	if(filename) // If none is specified, then go to look for next_map.json, for map rotation purposes.
-
-		//Default to MAP_DIRECTORY_MAPS if no directory is passed
-		if(directory)
-			if(!(directory in MAP_DIRECTORY_WHITELIST))
-				log_world("map directory not in whitelist: [directory] for map [filename]")
-				return config
-		else
-			directory = MAP_DIRECTORY_MAPS
-
-		filename = "[directory]/[filename].json"
-	else
-		filename = PATH_TO_NEXT_MAP_JSON
-
-
+/proc/load_map_config(filename = "data/next_map.json", default_to_box, delete_after, error_if_missing = TRUE)
+	var/datum/map_config/config = new
+	if (default_to_box)
+		return config
 	if (!config.LoadConfig(filename, error_if_missing))
 		qdel(config)
-		return load_default_map_config()
+		config = new /datum/map_config  // Fall back to Box
+	if (delete_after)
+		fdel(filename)
 	return config
 
-
 #define CHECK_EXISTS(X) if(!istext(json[X])) { log_world("[##X] missing from json!"); return; }
-
 /datum/map_config/proc/LoadConfig(filename, error_if_missing)
 	if(!fexists(filename))
 		if(error_if_missing)
@@ -128,46 +87,18 @@
 		log_world("map_file missing from json!")
 		return
 
-	if (islist(json["shuttles"]))
-		var/list/L = json["shuttles"]
-		for(var/key in L)
-			var/value = L[key]
-			shuttles[key] = value
-	else if ("shuttles" in json)
-		log_world("map_config shuttles is not a list!")
-		return
-
 	traits = json["traits"]
 	// "traits": [{"Linkage": "Cross"}, {"Space Ruins": true}]
 	if (islist(traits))
 		// "Station" is set by default, but it's assumed if you're setting
 		// traits you want to customize which level is cross-linked
 		for (var/level in traits)
-			if (!(ZTRAIT_STATION in level))
-				level[ZTRAIT_STATION] = TRUE
+			if (!(ZTRAIT_FORTRESS in level))
+				level[ZTRAIT_FORTRESS] = TRUE
 	// "traits": null or absent -> default
 	else if (!isnull(traits))
 		log_world("map_config traits is not a list!")
 		return
-
-	var/temp = json["space_ruin_levels"]
-	if (isnum(temp))
-		space_ruin_levels = temp
-	else if (!isnull(temp))
-		log_world("map_config space_ruin_levels is not a number!")
-		return
-
-	temp = json["space_empty_levels"]
-	if (isnum(temp))
-		space_empty_levels = temp
-	else if (!isnull(temp))
-		log_world("map_config space_empty_levels is not a number!")
-		return
-
-	if ("minetype" in json)
-		minetype = json["minetype"]
-
-	allow_custom_shuttles = json["allow_custom_shuttles"] != FALSE
 
 	if ("job_changes" in json)
 		if(!islist(json["job_changes"]))
@@ -187,4 +118,4 @@
 		. += "_maps/[map_path]/[file]"
 
 /datum/map_config/proc/MakeNextMap()
-	return config_filename == PATH_TO_NEXT_MAP_JSON || fcopy(config_filename, PATH_TO_NEXT_MAP_JSON)
+	return config_filename == "data/next_map.json" || fcopy(config_filename, "data/next_map.json")
